Tests: Remove old GFAL2 and WebDAV Tests; Fix #5962
Removed the "_new" suffix from the newly written test files, thus
finally replacing the old test_rse_protocols with the new ones. Also
changed the names of the files called within integration_tests.yml.

test_rse_protocol_gfal2_new.py -> test_rse_protocol_gfal2.py
test_rse_protocol_webdav_new.py -> test_rse_protocol_webdav.py
